Rats model

Run Rats data model

rats_example()

Value

stanfit object

See also

Examples

set.seed(123)
rats <- rats_example()

#> rstan (Version 2.18.2, GitRev: 2e1f913d3ca3)

#> For execution on a local, multicore CPU with excess RAM we recommend calling
#> options(mc.cores = parallel::detectCores()).
#> To avoid recompilation of unchanged Stan programs, we recommend calling
#> rstan_options(auto_write = TRUE)

rats

#> Inference for Stan model: rats.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean    sd    2.5%     25%     50%     75%   97.5%
#> alpha[1]       240.00    0.04  2.73  234.56  238.19  240.03  241.81  245.40
#> alpha[2]       247.87    0.04  2.70  242.58  246.07  247.87  249.64  253.27
#> alpha[3]       252.35    0.04  2.70  246.88  250.56  252.34  254.12  257.70
#> alpha[4]       232.57    0.04  2.73  227.08  230.81  232.53  234.42  238.04
#> alpha[5]       231.70    0.04  2.69  226.41  229.84  231.68  233.50  236.98
#> alpha[6]       249.67    0.03  2.69  244.36  247.87  249.65  251.51  255.10
#> alpha[7]       228.84    0.04  2.74  223.33  227.09  228.81  230.62  234.33
#> alpha[8]       248.21    0.04  2.64  243.12  246.38  248.25  250.05  253.18
#> alpha[9]       283.05    0.04  2.75  277.57  281.25  283.08  284.86  288.48
#> alpha[10]      219.46    0.04  2.71  214.11  217.64  219.47  221.21  224.76
#> alpha[11]      258.13    0.04  2.73  252.92  256.29  258.07  259.96  263.41
#> alpha[12]      228.26    0.04  2.70  222.80  226.50  228.25  230.05  233.73
#> alpha[13]      242.40    0.04  2.79  236.77  240.59  242.50  244.26  247.84
#> alpha[14]      268.12    0.04  2.67  262.95  266.33  268.14  269.93  273.31
#> alpha[15]      242.85    0.03  2.67  237.56  241.05  242.83  244.64  248.04
#> alpha[16]      245.25    0.04  2.66  239.96  243.49  245.22  246.99  250.52
#> alpha[17]      232.26    0.04  2.73  227.08  230.41  232.24  234.11  237.64
#> alpha[18]      240.54    0.04  2.72  235.19  238.74  240.53  242.40  245.92
#> alpha[19]      253.72    0.04  2.71  248.45  251.92  253.68  255.61  259.11
#> alpha[20]      241.65    0.04  2.69  236.32  239.84  241.62  243.47  246.82
#> alpha[21]      248.50    0.03  2.64  243.14  246.81  248.52  250.26  253.71
#> alpha[22]      225.34    0.04  2.76  219.89  223.55  225.35  227.14  230.78
#> alpha[23]      228.63    0.04  2.69  223.47  226.80  228.57  230.54  233.75
#> alpha[24]      245.12    0.04  2.79  239.54  243.28  245.09  247.00  250.63
#> alpha[25]      234.52    0.04  2.66  229.32  232.74  234.50  236.30  239.79
#> alpha[26]      253.95    0.04  2.71  248.69  252.10  253.97  255.74  259.20
#> alpha[27]      254.29    0.04  2.67  249.11  252.45  254.33  256.14  259.47
#> alpha[28]      243.03    0.04  2.77  237.65  241.17  243.09  244.86  248.43
#> alpha[29]      218.10    0.04  2.75  212.60  216.25  218.13  219.94  223.58
#> alpha[30]      241.46    0.04  2.72  236.07  239.59  241.50  243.29  246.73
#> beta[1]          6.06    0.00  0.25    5.59    5.90    6.06    6.23    6.56
#> beta[2]          7.07    0.00  0.27    6.55    6.89    7.07    7.24    7.58
#> beta[3]          6.48    0.00  0.24    6.00    6.32    6.49    6.64    6.97
#> beta[4]          5.33    0.00  0.26    4.81    5.16    5.33    5.50    5.85
#> beta[5]          6.58    0.00  0.25    6.09    6.41    6.58    6.74    7.05
#> beta[6]          6.18    0.00  0.25    5.68    6.01    6.17    6.35    6.67
#> beta[7]          5.97    0.00  0.25    5.48    5.81    5.97    6.14    6.47
#> beta[8]          6.42    0.00  0.25    5.93    6.26    6.42    6.58    6.92
#> beta[9]          7.07    0.00  0.27    6.55    6.90    7.07    7.25    7.58
#> beta[10]         5.85    0.00  0.24    5.38    5.68    5.84    6.01    6.31
#> beta[11]         6.81    0.00  0.26    6.31    6.64    6.81    6.99    7.30
#> beta[12]         6.11    0.00  0.25    5.62    5.95    6.11    6.27    6.61
#> beta[13]         6.16    0.00  0.25    5.68    5.99    6.16    6.33    6.65
#> beta[14]         6.70    0.00  0.26    6.19    6.52    6.69    6.87    7.21
#> beta[15]         5.41    0.00  0.26    4.92    5.22    5.41    5.59    5.91
#> beta[16]         5.92    0.00  0.25    5.43    5.76    5.92    6.08    6.41
#> beta[17]         6.28    0.00  0.25    5.77    6.11    6.28    6.45    6.78
#> beta[18]         5.84    0.00  0.25    5.35    5.67    5.84    6.00    6.32
#> beta[19]         6.41    0.00  0.25    5.93    6.24    6.41    6.58    6.90
#> beta[20]         6.06    0.00  0.25    5.56    5.89    6.06    6.22    6.54
#> beta[21]         6.41    0.00  0.25    5.91    6.25    6.41    6.57    6.91
#> beta[22]         5.85    0.00  0.26    5.36    5.68    5.85    6.03    6.35
#> beta[23]         5.74    0.00  0.25    5.25    5.58    5.74    5.91    6.22
#> beta[24]         5.88    0.00  0.24    5.40    5.72    5.88    6.05    6.37
#> beta[25]         6.92    0.00  0.26    6.41    6.75    6.92    7.10    7.43
#> beta[26]         6.56    0.00  0.25    6.07    6.39    6.56    6.73    7.04
#> beta[27]         5.90    0.00  0.25    5.42    5.73    5.89    6.07    6.39
#> beta[28]         5.84    0.00  0.25    5.35    5.66    5.84    6.02    6.33
#> beta[29]         5.66    0.00  0.25    5.18    5.49    5.66    5.82    6.15
#> beta[30]         6.13    0.00  0.25    5.63    5.97    6.13    6.30    6.63
#> mu_alpha       242.56    0.04  2.53  237.66  240.86  242.52  244.30  247.52
#> mu_beta          6.19    0.00  0.12    5.96    6.11    6.19    6.26    6.42
#> sigmasq_y       38.06    0.13  5.90   28.28   33.87   37.40   41.54   51.41
#> sigmasq_alpha  180.47    0.57 38.88  116.59  152.83  176.34  203.98  266.54
#> sigmasq_beta     0.31    0.00  0.11    0.14    0.23    0.29    0.36    0.58
#> sigma_y          6.15    0.01  0.47    5.32    5.82    6.12    6.45    7.17
#> sigma_alpha     13.36    0.02  1.43   10.80   12.36   13.28   14.28   16.33
#> sigma_beta       0.55    0.00  0.10    0.38    0.48    0.54    0.60    0.76
#> alpha0         106.42    0.05  3.68   99.29  104.00  106.37  108.94  113.67
#> lp__          -433.74    0.22  7.06 -448.41 -438.42 -433.37 -428.84 -421.23
#>               n_eff Rhat
#> alpha[1]       5310    1
#> alpha[2]       4683    1
#> alpha[3]       5144    1
#> alpha[4]       6023    1
#> alpha[5]       5385    1
#> alpha[6]       5983    1
#> alpha[7]       5510    1
#> alpha[8]       4035    1
#> alpha[9]       5993    1
#> alpha[10]      5396    1
#> alpha[11]      5516    1
#> alpha[12]      5180    1
#> alpha[13]      6184    1
#> alpha[14]      5337    1
#> alpha[15]      6176    1
#> alpha[16]      4899    1
#> alpha[17]      4563    1
#> alpha[18]      5540    1
#> alpha[19]      5226    1
#> alpha[20]      4798    1
#> alpha[21]      5763    1
#> alpha[22]      5345    1
#> alpha[23]      5691    1
#> alpha[24]      5786    1
#> alpha[25]      4800    1
#> alpha[26]      5666    1
#> alpha[27]      5531    1
#> alpha[28]      5498    1
#> alpha[29]      5277    1
#> alpha[30]      5918    1
#> beta[1]        4995    1
#> beta[2]        5151    1
#> beta[3]        5380    1
#> beta[4]        4404    1
#> beta[5]        4907    1
#> beta[6]        5365    1
#> beta[7]        5202    1
#> beta[8]        5115    1
#> beta[9]        5152    1
#> beta[10]       5109    1
#> beta[11]       5783    1
#> beta[12]       4410    1
#> beta[13]       5414    1
#> beta[14]       4774    1
#> beta[15]       4205    1
#> beta[16]       5088    1
#> beta[17]       5476    1
#> beta[18]       5360    1
#> beta[19]       5068    1
#> beta[20]       5039    1
#> beta[21]       5448    1
#> beta[22]       5177    1
#> beta[23]       5679    1
#> beta[24]       5389    1
#> beta[25]       4674    1
#> beta[26]       5504    1
#> beta[27]       5161    1
#> beta[28]       5232    1
#> beta[29]       4836    1
#> beta[30]       5346    1
#> mu_alpha       5003    1
#> mu_beta        4027    1
#> sigmasq_y      1939    1
#> sigmasq_alpha  4608    1
#> sigmasq_beta   3212    1
#> sigma_y        1965    1
#> sigma_alpha    4707    1
#> sigma_beta     3091    1
#> alpha0         4635    1
#> lp__           1003    1
#> 
#> Samples were drawn using NUTS(diag_e) at Tue May 14 22:58:25 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).