Add covrpage to Travis CI for a git directory

Adds bash file to execute after Travis test is successfully completed.

use_covrpage(file = ".travis/covrpage.sh", travis_type = c("bash",
  "pkgdown"), travis_blocks = c("after_success", "after_failure",
  "deploy_ghpages"), gh_user = gsub("@(.*?)$", "",
  system("git config user.email", intern = TRUE)), push_branch = "test",
  repo = system("git config travis.slug", intern = TRUE),
  deploy_branch = system("git rev-parse --abbrev-ref HEAD", intern = TRUE))

Arguments

file
character, file path to save bash script to, Default: '.travis/covrpage.sh'

travis_type
character, build .travis.yml template for bash or pkgdown deployment, Default: c('bash','pkgdown')

travis_blocks
character, vector that defines what .travis.yml blocks tocreate, Default: c('after_success','after_failure','deploy_ghpages')

gh_user
character, Github user name, Default: gsub("@(.*?)$", "", system("git config user.email", intern = TRUE))

push_branch
character, branch Travis will create to push updated covrpage readme file, Default: 'test'

repo
character, name of git repo that package is hosted on, Default: system("git config travis.slug", intern = TRUE)

deploy_branch
character, name of branch that current package build is located, Default: system("git rev-parse --abbrev-ref HEAD", intern = TRUE) (usually 'master')

Examples

use_covrpage(travis_type = 'bash',file=file.path(tempdir(),'myfile.sh'))

#> Writing myfile.sh bash script to /var/folders/kx/t4h_mm1910sb7vhm_gnfnx2c0000gn/T//RtmpxG3p46

#> Add to .travs.yml 
#> 
#>   r_github_packages: 
#>   - yonicd/covrpage
#>   
#>     after_sucess: 
#>   - Rscript -e 'covr::codecov()'
#>   - bash /var/folders/kx/t4h_mm1910sb7vhm_gnfnx2c0000gn/T//RtmpxG3p46/myfile.sh
#>     after_failure:
#>   - bash /var/folders/kx/t4h_mm1910sb7vhm_gnfnx2c0000gn/T//RtmpxG3p46/myfile.sh
#>   deploy:
#>     provider: pages
#>     skip-cleanup: true
#>     github-token: $GH_PAT
#>     keep-history: true
#>     local-dir: docs
#>     on:
#>       branch: master

#does not write to disk
use_covrpage(travis_type = 'bash',file=NULL)

#> #!/bin/bash
#>   
#> set -x
#> if [ $TRAVIS_BRANCH == "master" ] ; then
#> 
#> git config --global user.email "travis@travis-ci.org"
#> git config --global user.name "Travis CI"
#> 
#> git checkout -b test
#> 
#> Rscript -e "devtools::install() ; covrpage::covrpage_ci()"
#> 
#> git add .
#> git commit --message "Travis build: $TRAVIS_BUILD_NUMBER [skip ci]"
#> 
#> git remote add deploy https://yonis:${GH_PAT}@github.com/yonicd/covrpage.git
#> git push -f deploy test -v
#> 
#> else
#> echo "Not deploying, since this branch is not master."
#> fi

use_covrpage(travis_type = 'pkgdown')

#> Add to .travs.yml 
#> 
#>   r_github_packages: 
#>   - yonicd/covrpage   
#>   - r-lib/pkgdown
#> 
#>   after_success:
#>   - Rscript -e 'covr::codecov()'
#>   - Rscript -e 'devtools::install(); covrpage::covrpage_ci()'
#>   - Rscript -e 'pkgdown::build_site()'
#>   after_failure:
#>   - Rscript -e 'devtools::install(); covrpage::covrpage_ci()'
#>   - Rscript -e 'pkgdown::build_site()'
#>   deploy:
#>     provider: pages
#>     skip-cleanup: true
#>     github-token: $GH_PAT
#>     keep-history: true
#>     local-dir: docs
#>     on:
#>       branch: master

use_covrpage(travis_type = 'pkgdown',travis_block = 'after_success')

#> Add to .travs.yml 
#> 
#>   r_github_packages: 
#>   - yonicd/covrpage   
#>   - r-lib/pkgdown
#> 
#>   after_success:
#>   - Rscript -e 'covr::codecov()'
#>   - Rscript -e 'devtools::install(); covrpage::covrpage_ci()'
#>   - Rscript -e 'pkgdown::build_site()'
#>   
#>   

use_covrpage(travis_type = 'pkgdown',travis_block = 'after_failure')

#> Add to .travs.yml 
#> 
#>   r_github_packages: 
#>   - yonicd/covrpage   
#>   - r-lib/pkgdown
#> 
#>   
#>   after_failure:
#>   - Rscript -e 'devtools::install(); covrpage::covrpage_ci()'
#>   - Rscript -e 'pkgdown::build_site()'
#>   

use_covrpage(travis_type = 'pkgdown',travis_block = c('after_success','after_failure'))

#> Add to .travs.yml 
#> 
#>   r_github_packages: 
#>   - yonicd/covrpage   
#>   - r-lib/pkgdown
#> 
#>   after_success:
#>   - Rscript -e 'covr::codecov()'
#>   - Rscript -e 'devtools::install(); covrpage::covrpage_ci()'
#>   - Rscript -e 'pkgdown::build_site()'
#>   after_failure:
#>   - Rscript -e 'devtools::install(); covrpage::covrpage_ci()'
#>   - Rscript -e 'pkgdown::build_site()'
#>